core(playwright):Install and configure playwright
