+++ docs/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test-ct": "playwright test -c playwright-ct.config.ts"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.3",
@@ -20,6 +21,7 @@
     "react-router-dom": "^6.4.2"
   },
   "devDependencies": {
+    "@playwright/experimental-ct-react": "^1.29.1",
     "@tailwindcss/forms": "^0.5.3",
     "@testing-library/react": "^13.4.0",
     "@types/react": "^18.0.21",
