+++ docs/playwright/index.ts
@@ -1,0 +1,2 @@
+// Import styles, initialize component theme here.
+// import '../src/common.css';
