+++ docs/src/components/Loading/Loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Loading: React.FC = () => {
+  return (
+    <div>
+      <p>Por favor Espere...</p>
+    </div>
+  );
+};
+export default Loading;
