+++ docs/src/main.tsx
@@ -1,13 +1,16 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import './index.css';
+import { BrowserRouter as Router } from 'react-router-dom';
+import Loading from './components/Loading/Loading';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </React.StrictMode>
+      <Router>
+          <Suspense fallback={<Loading />}>
+            <App />
+          </Suspense>
+      </Router>
+  </React.StrictMode>,
 );
